+++ src/components/FhiToast/FhiToast.tsx
@@ -6,38 +6,8 @@ const MarginToast = styled(Toast)`
   margin-bottom: 1rem;
 `;
 
-const ToastBody = styled(Toast.Body)`
-  display: flex;
-  justify-content: space-between;
-  color: ${({ theme }) => theme.tekst};
-  padding: 20px;
-  border-radius: 0.25rem;
-  box-shadow: ${({ theme }) => `0 0.25rem 0.5rem ${theme.fhiGrey1}`};
-  ${({ type, theme }) => {
-    switch (type) {
-      case 'success':
-        return `background: ${theme.toastSuccess || theme.fhiGreen1}`;
-      case 'warning':
-        return `background: ${theme.toastWarning || theme.fhiOrange1}`;
-      case 'error':
-        return `background: ${theme.toastError || theme.fhiPinkLight3}`;
-      default:
-        return `background: ${theme.toastNetural || theme.fhiBlueGrey2}`;
-    }
-  }}
-`;
-
-const CloseButton = styled.button`
-  align-self: start;
-  margin-top: -1.5rem;
-  margin-right: -1rem;
-  font-size: 2rem;
-  min-width: 2rem;
-  border: none;
-  background: none;
-`;
-
 type FhiToastProps = {
+  title?: string,
   text: string,
   type?: 'success' | 'warning' | 'error',
   delay?: number,
@@ -45,6 +15,7 @@ type FhiToastProps = {
 };
 
 const FhiToast:FC<FhiToastProps> = ({
+  title,
   text,
   type,
   delay,
@@ -53,22 +24,21 @@ const FhiToast:FC<FhiToastProps> = ({
   <MarginToast
     onClose={removeToast}
     delay={delay}
-    autohide
+    bg={type}
+    autohide={delay !== 0}
   >
-    <ToastBody
-      type={type}
-    >
+    <Toast.Header>
+      <strong className="me-auto">{title}</strong>
+    </Toast.Header>
+    <Toast.Body>
       {text}
-      <CloseButton type="button" className="btn-close ml-2 mb-1" onClick={removeToast}>
-        <span aria-hidden="true">×</span>
-        <span className="sr-only">Close</span>
-      </CloseButton>
-    </ToastBody>
+    </Toast.Body>
   </MarginToast>
 );
 
 FhiToast.defaultProps = {
   type: undefined,
+  title: undefined,
   delay: 5000,
 };