+++ CHANGELOG.md
@@ -10,6 +10,11 @@
 * :nut_and_bolt: **New** Added `flush` and `alternate` variations to `FhiAccordion`
 * :nut_and_bolt: **New** Added `FhiTabs` with new design
 * :boom: **Breaking** Replaced `FhiTabBar` with `FhiTabs`
+* :boom: **Breaking** Updated design on `FhiToast`
+* :nut_and_bolt: **New** Added `title` to `FhiToast`
+* :bug: **Bugfix** Fixed delay for `FhiToast` (Added option for 0 delay that makes toast stay until closed by user)
+
+
 
 ## 1.3.0